moved socket error handler to client class
