+++ client/src/easel/easelManager.ts
@@ -344,8 +344,9 @@ export class EaselManager {
     });
     this.currentStageWidth = width; // todo maybe put this in one nice object
     this.updateEaselScale();
-    document.querySelector("body").style.backgroundImage =
-      `url(${backgroundRoute}forest.png)`;
+    document.querySelector(
+      "body"
+    ).style.backgroundImage = `url(${backgroundRoute}forest.png)`;
     this.easelDiv.appendChild(this.layoutPivot);
   };
 
@@ -388,13 +389,6 @@ export class EaselManager {
         this.constructStage(stage);
       }
     );
-    client.on("socketError", "easel", () => {
-      const errorImg = document.createElement("img");
-      errorImg.src = `${iconRoute}networkError.png`;
-      errorImg.alt = "Network error occured!";
-      errorImg.classList.add("network-error");
-      document.body.appendChild(errorImg);
-    });
 
     this.easelDiv.style.setProperty(
       "--easel__prop-sprite--border-color",
